shift ShipsComputer to using Longs for day 9 part 1

+++ src/main/kotlin/day2/ShipsComputer.kt
@@ -152,5 +152,5 @@ class ShipsComputer {
         }
     }
 
-    private fun readOpCode(pointer: Int): OpCodeInstruction = OpCodeInstruction(memory[pointer])
+    private fun readOpCode(pointer: Int): OpCodeInstruction = OpCodeInstruction(memory[pointer].toLong())
 }
