+++ src/test/kotlin/day5/UpgradeTheShipsComputerWithChannels.kt
@@ -16,8 +16,8 @@ class UpgradeTheShipsComputerWithChannels {
     fun `OpCode 4 takes a value from an index and writes to the program output`() {
         val program = "4,3,99,201"
 
-        val inputs = Channel<Int>()
-        val outputs = Channel<Int>()
+        val inputs = Channel<Long>()
+        val outputs = Channel<Long>()
         val halts = Channel<Boolean>()
 
         val computer = ShipsComputerWithChannels(
@@ -43,8 +43,8 @@ class UpgradeTheShipsComputerWithChannels {
     fun `3,0,4,0,99 takes whatever it is given as input and returns it as output`() {
         val program = "3,0,4,0,99"
 
-        val inputs = Channel<Int>()
-        val outputs = Channel<Int>()
+        val inputs = Channel<Long>()
+        val outputs = Channel<Long>()
         val halts = Channel<Boolean>()
 
         val computer = ShipsComputerWithChannels(
@@ -59,7 +59,7 @@ class UpgradeTheShipsComputerWithChannels {
         }
 
         runBlocking {
-            val input = Random.Default.nextInt()
+            val input = Random.Default.nextLong()
             inputs.send(input)
             assertThat(outputs.receive()).isEqualTo(input)
             inputs.close()
@@ -72,8 +72,8 @@ class UpgradeTheShipsComputerWithChannels {
     fun `part 1 solution`() {
         val program = this::class.java.getResource("/day5/puzzleInput.txt").readText().trim()
 
-        val inputs = Channel<Int>()
-        val outputs = Channel<Int>()
+        val inputs = Channel<Long>()
+        val outputs = Channel<Long>()
         val halts = Channel<Boolean>()
 
         val computer = ShipsComputerWithChannels(
@@ -88,9 +88,9 @@ class UpgradeTheShipsComputerWithChannels {
         }
 
         runBlocking {
-            inputs.send(1)
+            inputs.send(1L)
             for (output in outputs) {
-                if (output != 0) {
+                if (output != 0L) {
                     assertThat(output).isEqualTo(15314507)
                     inputs.close()
                     outputs.close()
